+++ src/ipa/simple/black_level.h
@@ -2,12 +2,13 @@
 /*
  * Copyright (C) 2024, Red Hat Inc.
  *
- * black_level.h - black level handling
+ * black level handling
  */
 
 #pragma once
 
 #include <array>
+#include <stdint.h>
 
 #include "libcamera/internal/software_isp/swisp_stats.h"
 
@@ -17,11 +18,11 @@ class BlackLevel
 {
 public:
 	BlackLevel();
-	unsigned int get() const;
+	uint8_t get() const;
 	void update(SwIspStats::Histogram &yHistogram);
 
 private:
-	unsigned int blackLevel_;
+	uint8_t blackLevel_;
 	bool blackLevelSet_;
 };
